format code

+++ product-query-service/src/main/java/com/javatechie/dto/FilterProductReq.java
@@ -1,11 +1,12 @@
 package com.javatechie.dto;
 
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record FilterProductReq(String name,
-                               String description,
-                               @JsonProperty("price_from") Integer priceFrom,
-                               @JsonProperty("price_to") Integer priceTo,
-                               int page,
-                               int size) {
-}
+public record FilterProductReq(
+        String name,
+        String description,
+        @JsonProperty("price_from") Double priceFrom,
+        @JsonProperty("price_to") Double priceTo,
+        int page,
+        int size) {}
